+++ Tasks/LambdaNETCoreDeploy/helpers/NetCoreDeployTaskOperations.ts
@@ -17,11 +17,6 @@ export class TaskOperations {
 
     public static async deployFunction(taskParameters: Parameters.TaskParameters): Promise<void> {
 
-        // assume role credentials are not currently supported
-        if (taskParameters.AssumeRoleARN) {
-            throw new Error('Assume role-based temporary credentials are not currently supported for this task');
-        }
-
         const cwd = this.determineProjectDirectory(taskParameters.lambdaProjectPath);
         console.log(tl.loc('DeployingProjectAt', cwd));
 
@@ -39,14 +34,18 @@ export class TaskOperations {
             }
         }
 
-        // temp code until we adopt support for assume role credentials,
-        // at which point we'll switch to using taskParameters.Credentials
-        const awsEndpoint = tl.getInput('awsCredentials', true);
-        const awsEndpointAuth = tl.getEndpointAuthorization(awsEndpoint, false);
-
         const env = process.env;
-        env.AWS_ACCESS_KEY_ID = awsEndpointAuth.parameters.username;
-        env.AWS_SECRET_ACCESS_KEY = awsEndpointAuth.parameters.password;
+        // if assume role credentials are in play, make sure the initial generation
+        // of temporary credentials has been performed
+        await taskParameters.Credentials.getPromise().then(() => {
+            env.AWS_ACCESS_KEY_ID = taskParameters.Credentials.accessKeyId;
+            env.AWS_SECRET_ACCESS_KEY = taskParameters.Credentials.secretAccessKey;
+            if (taskParameters.Credentials.sessionToken) {
+                env.AWS_SESSION_TOKEN = taskParameters.Credentials.sessionToken;
+            }
+        });
+
+        await taskParameters.configureHttpProxyFromAgentProxyConfiguration('LambdaNETCoreDeploy');
 
         const wrapper = new DotNetCliWrapper(cwd, env);